gui/List: precise MouseWheel traversing.

+++ include/input/InputManager.h
@@ -118,7 +118,7 @@ namespace ProjectSpace
 		sf::Vector2i getMousePosition() const;
 
 		// 0 is not moved.
-		int getMouseWheelDelta() const;
+		float getMouseWheelDelta() const;
 
 		void registerInputContext(std::string const& name, InputContext* inputContext);
 		void blockInputContext(std::string const& name);
@@ -137,7 +137,7 @@ namespace ProjectSpace
 		sf::Mouse::Button lastUpdatetMouseButton;
 		bool mouseMovedThisFrame;
 		sf::Vector2i currentMousePosition;
-		int mouseWheelDelta;
+		float mouseWheelDelta;
 
 		// TODO: std::map<std::string, std::pair<InputContext*, bool>> inputContexts.
 		// The std::set is not necessary for checking blocked InputContexts.
